+++ src/main.tsx
@@ -4,6 +4,7 @@ import { StyledEngineProvider } from '@mui/material/styles';
 import {createRoot} from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary.tsx';
 import { AudioProvider } from './contexts/AudioContext.tsx';
 import { AuthProvider } from './contexts/AuthContext.tsx';
 import { AppThemeProvider } from './theme/AppThemeProvider';
@@ -11,17 +12,19 @@ import './index.css';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <StyledEngineProvider enableCssLayer>
-      <GlobalStyles styles="@layer theme, base, mui, components, utilities;" />
-      <AppThemeProvider>
-        <BrowserRouter>
-          <AuthProvider>
-            <AudioProvider>
-              <App />
-            </AudioProvider>
-          </AuthProvider>
-        </BrowserRouter>
-      </AppThemeProvider>
-    </StyledEngineProvider>
+    <ErrorBoundary>
+      <StyledEngineProvider enableCssLayer>
+        <GlobalStyles styles="@layer theme, base, mui, components, utilities;" />
+        <AppThemeProvider>
+          <BrowserRouter>
+            <AuthProvider>
+              <AudioProvider>
+                <App />
+              </AudioProvider>
+            </AuthProvider>
+          </BrowserRouter>
+        </AppThemeProvider>
+      </StyledEngineProvider>
+    </ErrorBoundary>
   </StrictMode>,
 );